Move link template expansion logic to URITemplate
